Created Utility.cpp and Utility.h Isolated and moved read_error to Utility (alongside with print_logo) Created config file Options are now read from "config.txt" fixed some line spacing
