+++ Utility.h
@@ -1,0 +1,10 @@
+#ifndef UTILITY
+#define UTILITY
+
+#include <string>
+
+void print_logo();
+
+bool read_error(std::string filename);
+
+#endif
